Fix for debouncing not working due
Debouncer was using a uint to store last valid interrupt time,
however the time was ulong, so that mean truncation and false positives
during the comparisons.

Added a debouncer class.

+++ Firmware/ArduinoEmptyFunctions.cpp
@@ -2,7 +2,7 @@
 
 void delay(int x) {}
 void delayMicroseconds(int x) {}
-int millis() { return 0; }
+unsigned long millis() { return 0; }
 
 void pinMode(int x, int y) {}
 void digitalWrite(int x, int y) {}
